chore: Move Batch execute setting to appearance key in store
